+++ src/screens/AddMood/index.js
@@ -8,6 +8,7 @@ import sad from '../../assets/humores/sad.png';
 import terrible from '../../assets/humores/terrible.png';
 import { SafeAreaView } from 'react-native-safe-area-context';
 import { useNavigation } from '@react-navigation/native';
+import style from './style';
 
 const AddMood = () => {
     const Navigation = useNavigation()
@@ -32,12 +33,10 @@ const AddMood = () => {
         }
     }
     return (
-        
         <ScrollView>
             <SafeAreaView style={style.tresLinhas}>
                 <TouchableOpacity style={style.bottomClose}
-                onPress={() => Navigation.goBack()}
-                >
+                    onPress={() => Navigation.goBack()}>
                     <Icon name='close'
                         size={33}
                         color='#697FFF' />
@@ -194,175 +193,4 @@ const AddMood = () => {
         </ScrollView>
     )
 }
-const style = StyleSheet.create({
-    tresLinhas: {
-        flex: 1,
-        backgroundColor: 'white',
-        alignItems: 'center'
-    },
-    TextHow: {
-        fontWeight: '800',
-        color: 'black',
-        fontSize: 23,
-        fontFamily: 'SourceSansPro-Black',
-        marginBottom: 9
-    },
-    IconSelect: {
-        borderRadius: 20,
-        backgroundColor: 'white',
-        marginTop: 10,
-        padding: 10,
-        width: '90%',
-        height: 450,
-        borderWidth: 1,
-        justifyContent: 'space-evenly',
-    },
-    date: {
-        width: '60%',
-        flexDirection: 'row',
-        justifyContent: 'space-evenly',
-    },
-    TextDate: {
-        fontSize: 13,
-        color: '#969696',
-        fontWeight: '500',
-        fontFamily: 'SourceSansPro-SemiBold'
-    },
-    bottomClose: {
-        justifyContent: 'flex-start',
-        alignSelf: 'flex-start',
-        marginTop: 15,
-        marginLeft: 20,
-        backgroundColor: 'rgba(48, 79, 254, 0.1)',
-        borderRadius: 8,
-    },
-    HowStyle: {
-        flex: 1,
-        justifyContent: 'center',
-        alignItems: 'center',
-        marginTop: 35
-    },
-    humorEmote: {
-        width: 43,
-        alignItems: 'center',
-        flexDirection: 'row',
-        height: 43,
-    },
-    HumorChoose: {
-        width: '80%',
-        flexDirection: 'row',
-        justifyContent: 'space-evenly',
-        height: 150,
-        alignItems: 'center'
-    },
-    HumorName: {
-        alignItems: 'center',
-        flexDirection: 'row',
-        justifyContent: 'space-around',
-    },
-    HumorFont: {
-        fontSize: 10,
-        fontWeight: 'bold',
-        color: '#969696',
-        fontFamily: 'SourceSansPro-SemiBold'
-    },
-    day: {
-        flexDirection: 'row',
-        alignItems: 'center',
-        justifyContent: 'space-evenly',
-        width: '58%'
-    },
-    hour: {
-        flexDirection: 'row',
-        alignItems: 'center',
-        justifyContent: 'space-evenly',
-        width: '20%'
-    },
-    Happend: {
-        borderRadius: 22,
-        backgroundColor: 'white',
-        marginTop: 25,
-        padding: 10,
-        width: '90%',
-        height: 98,
-        borderWidth: 1,
-        marginBottom: 3
-    },
-    Digit: {
-        marginBottom: 40
-    },
-    bottomConfirm: {
-        borderRadius: 7,
-        backgroundColor: '#304ffe',
-        marginBottom: 35,
-        alignItems: 'center',
-        justifyContent: 'center',
-        margin: 20,
-        width: 380,
-        height: 52,
-    },
-    SaveColor: {
-        color: 'white',
-        fontSize: 16,
-        fontWeight: '900',
-        fontFamily: 'SourceSansPro-SemiBold'
-    },
-    IconsOne: {
-        alignItems: 'center',
-        flexDirection: 'row',
-        justifyContent: 'space-around',
-        width: '100%',
-    },
-    IconsTwo: {
-        alignItems: 'center',
-        flexDirection: 'row',
-        justifyContent: 'space-around',
-        width: '100%'
-    },
-    IconsThree: {
-        alignItems: 'center',
-        flexDirection: 'row',
-        justifyContent: 'space-around',
-        width: '100%'
-    },
-    IconsBorder: {
-        borderRadius: 35,
-        borderWidth: 1,
-        width: 65,
-        height: 65,
-        paddingLeft: 9,
-        paddingTop: 9,
-    },
-    Actions: {
-        fontWeight: '600',
-        color: 'black',
-        fontSize: 12,
-        textAlign: 'center',
-
-    },
-    iconsAlign: {
-        alignItems: 'center',
-        justifyContent: 'center',
-        borderRadius: 59,
-        width: 67,
-        height: 67,
-
-    },
-    IconsStyle: {
-        alignItems: 'center',
-        height: 80,
-        width: 60,
-        justifyContent: 'center'
-    },
-    IconsTop: {
-        alignItems: 'center',
-        height: 53,
-        width: 52,
-        justifyContent: 'center',
-        borderRadius: 52,
-    },
-    alignIcons: {
-        alignItems: 'center'
-    }
-})
 export default AddMood
