+++ src/screens/CardMood/styles.js
